requirements.txt
torch
transformers
ffmpeg-python
scikit-learn
numpy
opencv-python
